refactor: reorganize redux reducers folders
