+++ packages/frontend/src/redux/createReducers/combinedAccountReducers.js
@@ -1,4 +1,4 @@
-import tokensSlice from './slices/tokens';
+import tokensSlice from '../slices/tokens';
 
 export default () => ({
     [tokensSlice.name]: tokensSlice.reducer
